requirements.txt
aiogram==3.15
python-dotenv==0.21.1
fastapi==0.111.0
uvicorn==0.21.1
pydantic==2.7.0
pydantic-settings==2.3.3
aiohttp==3.9.5